Initial commit
This is the first commit of my personal homepage and blog. I don't have
any plans of yet to blog but I hope to eventually. In the mean time I've
commited here a fresh install of jekyll with a simple home page showing
some information about myself and contact links.